feat(rust): isolate and expose bi-diretional stream names
The specification for bi-directional streams leaves the specifics of
the producer/ consumer relationship open to implementations.  The
approach taken in this implementation is not ideal (as it burdains the
user to know _both_ stream names in the tx/rx pair), but will do for
now.

When creating a bi-directional stream between nodes A and B on node A
the given parameters must be: stream-A-B for the TX stream, and
stream-B-A for the RX stream.  On node B it is the inverse: stream-B-A
for the TX stream, and stream-A-B for the RX stream.